+++ src/shapes/text.js
@@ -37,7 +37,9 @@ export default function(settings: Object, _this: Global): GraghShape {
     const size = fontSize * scale;
     const font = `normal ${fontWeight} ${size}px ${fontFamily}`;
 
-    DefineScale.call(this, scale, 'x', 'y', 'width', 'height', 'moveX', 'moveY', 'paddingTop', 'paddingLeft');
+    if(!this.fixed) {
+      DefineScale.call(this, scale, 'x', 'y', 'width', 'height', 'moveX', 'moveY', 'paddingTop', 'paddingLeft');
+    }
 
     let textWidth, ellipsisText;
 
@@ -101,6 +103,8 @@ export default function(settings: Object, _this: Global): GraghShape {
     text: settings.text || 'no text',
     style: settings.style || 'fill',
     paddingTop: settings.paddingTop || 0,
-    paddingLeft: settings.paddingLeft || 0
+    paddingLeft: settings.paddingLeft || 0,
+    scaled_paddingTop: (settings.paddingTop || 0) * _this.scale,
+    scaled_paddingLeft: (settings.paddingLeft || 0) * _this.scale
   });
 }