+++ .github/CHANGELOG.md
@@ -96,6 +96,12 @@
 
 #### v4.0.0
 * 修复无法移除wheel事件的bug。
+* 添加api `setGlobalProps`， 用于设置OMG全局状态比如enableGlobalTranslate, enableGlobalScale。
+* Graph的config移除bg参数。
+* Graph的fixed属性升级，设置了fixed之后，全局拖拽，全局缩放，自身拖拽全部对其无效。
+* 添加页面重置api reset。
+* graph line支持拖拽
+* 添加finishAnimation接口。当omg中所有的动画结束后会执行这一函数。
 
 ## ToDo